wyoming-openwakeword/requirements.txt
tflite-runtime-nightly
wyoming==1.7.2
numpy<2
